Remove scorer.py and move its refactored logic to scoring_engine.py
